+++ src/components/forms/textfield/TextField.svelte
@@ -8,6 +8,7 @@
 
   export let label: string;
   export let value: string;
+  export let placeholder: string;
   export let id: string = `khao-ui-texfield-${label}`;
   export let type: TextFieldType = textFieldTypeDefault;
 </script>
@@ -16,9 +17,9 @@
   <label class="label" for={id}>{label}</label>
 
   {#if type === "output"}
-    <output class="input" {id}>{value}</output>
+    <output class="field" {id}>{value}</output>
   {:else}
-    <input class="input" {id} {type} {value} />
+    <input class="field" {placeholder} {id} {type} {value} />
   {/if}
 </div>